backend/requirements.txt
asgiref==3.2.7
certifi==2019.11.28
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==3.0.4
django-3-jet==1.0.8
django-admin-rangefilter==0.5.4
django-appconf==1.0.4
django-cacheops==4.2
django-cors-headers==3.2.1
django-dotenv==1.4.2
django-extensions==2.2.8
django-filter==2.2.0
django-ipware==2.1.0
django-leaflet==0.26.0
django-modeltranslation==0.14.4
django-redis==4.11.0
djangorestframework==3.11.0
djangorestframework-recaptcha==0.2.0
funcy==1.14
gunicorn==20.0.4
idna==2.9
itypes==1.1.0
Jinja2==2.11.1
Markdown==3.2.1
MarkupSafe==1.1.1
psycopg2==2.8.4
pytz==2019.3
PyYAML==5.3.1
redis==3.4.1
requests==2.23.0
six==1.14.0
sqlparse==0.3.1
uritemplate==3.0.1
urllib3==1.25.8
django-rest-auth==0.9.5